requirements.txt
SQLAlchemy==2.0.7
aiohttp==3.8.4
python-telegram-bot==20.2